requirements.txt
paho-mqtt==2.1.0
PyYAML==6.0.2
watchdog==5.0.2
prettytable==3.11.0
wcwidth=0.2.13